forms_advanced/requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2023.1
Django==4.2.9
django-crispy-forms==2.1
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.9.0
